+++ NEWS.md
@@ -1,6 +1,8 @@
 # capture 0.1.1
 
 * `capture()` / `capture_pdf()`: added `scale` argument, to be applied to image dimensions.
+* `capture()`: added `inputId` argument to allow retrieving image as base64 string server-side.
+
 
 # capture 0.1.0